Throw different dev sandbox components under tabs
