+++ packages/grafana-llm-app/src/components/AppConfig/DevSandbox/types.ts
@@ -1,0 +1,7 @@
+export interface RenderedToolCall {
+  name: string;
+  arguments: string;
+  running: boolean;
+  error?: string;
+  response?: any;
+}
